refactor[#345] : Region과 LngLat 타입을 entities/auth/api에서 가져오기
