+++ src/features/setting/api/postChangeRegion.ts
@@ -1,5 +1,5 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import type { Region } from "@/features/auth/api/region";
+import type { Region } from "@/entities/auth/api";
 import { useAuthStore } from "@/shared/store";
 import { SETTING_END_POINT } from "../constants";
